+++ vck-openid-ktor/src/commonMain/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VpWallet.kt
@@ -3,15 +3,17 @@ package at.asitplus.wallet.lib.ktor.openid
 import at.asitplus.KmmResult
 import at.asitplus.catching
 import at.asitplus.catchingUnwrapped
+import at.asitplus.dcapi.request.Oid4vpDCAPIRequest
 import at.asitplus.openid.AuthenticationRequestParameters
+import at.asitplus.openid.OpenIdConstants.Errors.INVALID_REQUEST
 import at.asitplus.openid.RelyingPartyMetadata
 import at.asitplus.openid.RequestParametersFrom
 import at.asitplus.wallet.lib.agent.HolderAgent
 import at.asitplus.wallet.lib.agent.KeyMaterial
 import at.asitplus.wallet.lib.data.CredentialPresentation
+import at.asitplus.wallet.lib.data.CredentialPresentationRequest
 import at.asitplus.wallet.lib.data.vckJsonSerializer
-import at.asitplus.dcapi.request.Oid4vpDCAPIRequest
-import at.asitplus.wallet.lib.data.MediaTypes
+import at.asitplus.wallet.lib.oidvci.OAuth2Error
 import at.asitplus.wallet.lib.oidvci.encodeToParameters
 import at.asitplus.wallet.lib.openid.AuthenticationResponseResult
 import at.asitplus.wallet.lib.openid.AuthorizationResponsePreparationState
@@ -23,7 +25,6 @@ import io.ktor.client.call.body
 import io.ktor.client.engine.HttpClientEngine
 import io.ktor.client.plugins.DefaultRequest
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.request.HttpRequestBuilder
 import io.ktor.client.request.forms.FormDataContent
 import io.ktor.client.request.forms.submitForm
 import io.ktor.client.request.get
@@ -34,8 +35,14 @@ import io.ktor.client.request.url
 import io.ktor.client.statement.HttpResponse
 import io.ktor.client.statement.bodyAsText
 import io.ktor.client.statement.readRawBytes
-import io.ktor.http.*
+import io.ktor.http.ContentType
+import io.ktor.http.HttpHeaders
+import io.ktor.http.HttpMethod
+import io.ktor.http.Parameters
+import io.ktor.http.URLBuilder
 import io.ktor.http.content.OutgoingContent
+import io.ktor.http.formUrlEncode
+import io.ktor.http.parameters
 import io.ktor.serialization.kotlinx.json.json
 import io.ktor.utils.io.core.toByteArray
 import kotlinx.coroutines.Dispatchers
@@ -62,7 +69,8 @@ class OpenId4VpWallet(
     sealed interface AuthenticationResult
 
     data class AuthenticationSuccess(val redirectUri: String? = null) : AuthenticationResult
-    data class AuthenticationForward(val authenticationResponseResult: AuthenticationResponseResult.DcApi) : AuthenticationResult
+    data class AuthenticationForward(val authenticationResponseResult: AuthenticationResponseResult.DcApi) :
+        AuthenticationResult
 
 
     private val client: HttpClient = HttpClient(engine) {
@@ -112,6 +120,21 @@ class OpenId4VpWallet(
     ): KmmResult<RequestParametersFrom<AuthenticationRequestParameters>> =
         openId4VpHolder.parseAuthenticationRequestParameters(input, dcApiRequest)
 
+    suspend fun createAuthnErrorResponse(
+        error: OAuth2Error,
+        request: RequestParametersFrom<AuthenticationRequestParameters>
+    ) = catchingUnwrapped {
+        Napier.i("createAuthnErrorResponse $error, $request")
+        openId4VpHolder.createAuthnErrorResponse(error = error, request = request).getOrThrow().let {
+            when (it) {
+                is AuthenticationResponseResult.Post -> postResponse(it)
+                is AuthenticationResponseResult.Redirect -> redirectResponse(it)
+                is AuthenticationResponseResult.DcApi ->
+                    throw UnsupportedOperationException("Unsupported error response mode")
+            }
+        }
+    }
+
     suspend fun startAuthorizationResponsePreparation(
         request: RequestParametersFrom<AuthenticationRequestParameters>
     ): KmmResult<AuthorizationResponsePreparationState> =
@@ -159,11 +182,20 @@ class OpenId4VpWallet(
             request = request,
             clientMetadata = clientMetadata,
             credentialPresentation = credentialPresentation
-        ).getOrThrow().let {
-            when (it) {
-                is AuthenticationResponseResult.Post -> postResponse(it)
-                is AuthenticationResponseResult.Redirect -> redirectResponse(it)
-                is AuthenticationResponseResult.DcApi -> AuthenticationForward(it)
+        ).getOrElse {
+            createAuthnErrorResponse(
+                error = OAuth2Error(
+                    error = INVALID_REQUEST,
+                    errorDescription = it.message,
+                    state = request.parameters.state
+                ), request = request
+            )
+            throw it
+        }.let { response ->
+            when (response) {
+                is AuthenticationResponseResult.Post -> postResponse(response)
+                is AuthenticationResponseResult.Redirect -> redirectResponse(response)
+                is AuthenticationResponseResult.DcApi -> AuthenticationForward(response)
             }
         }
     }
@@ -181,6 +213,22 @@ class OpenId4VpWallet(
             }
         )
     }
+
+    suspend fun getMatchingCredentials(
+        preparationState: AuthorizationResponsePreparationState,
+        request: RequestParametersFrom<AuthenticationRequestParameters>
+    ) = catchingUnwrapped {
+            openId4VpHolder.getMatchingCredentials(preparationState).getOrElse {
+                createAuthnErrorResponse(
+                    error = OAuth2Error(
+                        error = INVALID_REQUEST,
+                        errorDescription = it.message,
+                        state = request.parameters.state
+                    ), request = request
+                )
+                throw it
+            }
+        }
 
     /**
      * Our implementation of ktor's [FormDataContent], but with [contentType] without charset appended,